+++ AutoServiceManagment/AutoServiceManagment.DomainModels/DTOs/FinanceDto.cs
@@ -7,10 +7,10 @@ namespace AutoServiceManagment.DomainModels.DTOs
     public class FinanceDto : IDto
     {
         public int Id { get; set; }
-        public decimal CommunalCost { get; set; }
+        public double CommunalCost { get; set; }
         [Required(ErrorMessage = "Datetime is required!")]
 
         public DateTime Date { get; set; }
-        public decimal AdditionalCost { get; set; }
+        public double AdditionalCost { get; set; }
     }
 }
