+++ AutoServiceManagment/AutoServiceManagment.DomainModels/DTOs/SalaryDto.cs
@@ -16,9 +16,9 @@ namespace AutoServiceManagment.DomainModels.DTOs
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{mm/dd/yyyy}")]
         public DateTime Date { get; set; }
-        public decimal NetSalary { get; set; } = 0;
-        public decimal Bonus { get; set; }
-        public decimal Overtime { get; set; }
+        public double NetSalary { get; set; } = 0;
+        public double Bonus { get; set; }
+        public double Overtime { get; set; }
         [ForeignKey(nameof(Tax))]
         public int TaxID { get; set; }
         //public Tax Tax { get; set; }
