+++ AutoServiceManagment/AutoServiceManagment.DomainModels/DTOs/StatisticsDto.cs
@@ -7,7 +7,7 @@ namespace AutoServiceManagment.DomainModels.DTOs
     public class StatisticsDto : IDto
     {
         public int Id { get; set; }
-        public decimal Profit { get; set; }
+        public double Profit { get; set; }
 
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{mm/dd/yyyy}")]
